+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 4.0.0
+BREAKING CHANGES:
+* [Android] compileSdk has been upgraded to version 34.
+* [Android] Java version has been upgraded to version 17.
+
+## 3.5.7
+Improvements:
+* Updated js dependency together with other dependencies.
+* Reverted compileSdk version to 33 on Android. This update will be released under version 4.0.0.
+
+## 3.5.6
+Bugs fixed:
+* [web] Fixed a crash with the ZXing barcode format (thanks @hazzo!)
+* [web] Fixed stream controller not being closed on web.
+* [iOS] Fixed a crash with unsupported torch modes. (thanks @navaronbracke !)
+* [iOS] Fixed a crash with the camera discovery session. (thanks @navaronbracke !)
+
+Improvements:
+* Upgrade camera dependencies on Android.
+* Upgrade compileSdk version to 34 on Android.
+* Add numberOfCameras parameter in MobileScannerArguments callback, which shows how many cameras there are available on Android. 
+* [Android] Migrated to ResolutionSelector with ResolutionStrategy. You can opt in into the new selector by setting [useNewCameraSelector] in the [MobileScannerController] to true.
+
 ## 3.5.5
 Bugs fixed:
 * Fixed a bug where the scanner would get stuck after denying permissions on Android. (thanks @navaronbracke !)
@@ -37,6 +60,9 @@ Bugs fixed:
 * Fixed messages not being sent on the main thread for Android, iOS and MacOS. (thanks @navaronbracke !)
 
 ## 3.5.0
+
+**NOTE: From this version onwards, `mobile_scanner` requires Android projects to have a `compileSdk` of 34 (Android 14) or higher**
+
 New Features:
 * Added the option to switch between bundled and unbundled MLKit for Android. (thanks @woolfred !)
 * Added the option to specify the camera resolution for Android. (thanks @EArminjon !)